+++ app/src/main/res/layouts/diverse/layout/activity_generate_qrcode.xml
@@ -4,14 +4,17 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".QRCode.GenerateQRCode">
+    tools:context=".QRCode.GenerateQRCode"
+    android:background="@drawable/background">
 
     <Button
         android:id="@+id/generateCode"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="50dp"
-        android:text="Button"
+        android:text="Generează cod QR"
+        android:textSize="16sp"
+        android:background="@drawable/button"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
