+++ app/src/main/res/layouts/diverse/layout/add_medication.xml
@@ -6,8 +6,6 @@
     android:layout_height="match_parent"
     android:background="@drawable/background">
 
-
-
 <EditText
         android:id="@+id/txtDiagnostic"
         android:layout_width="match_parent"
